Fix ctxpipe project lon_0 bug; float32 coarse coords; tile plan (v0.10.1)
ctxpipe project was SIGKILL'd by macOS on any image whose longitude
was far from the prime meridian. Reproducer:
B17_016442_2252_XI_45N257W at lon~102°, lat 45°N — peak memory 66 GB
on a 24 GB machine, killed during compute_transform_coarse.

Root cause: _auto_projection() in ctxpipe/pipeline.py emitted
projection strings with `+lon_0=0.0` regardless of image location,
relying on a comment that was simply wrong ("will be overridden by
grid_from_params" — grid_from_params uses the CRS as-is). For a
meridional strip in Sinusoidal projection, the axis-aligned bounding
box inflates ~latitude_span × cos(center_lat) in X when lon_0 is far
from the image's actual longitude. Concretely, B17's 5°×1.7° strip
produced a 78,770×53,136 = 4.19-billion-pixel output instead of the
expected 12k×53k.

Fix: _auto_projection now takes lon_range and sets lon_0 to the mean
image longitude for both Sinusoidal and Polar Stereographic paths.
Caller in ctxpipe/pipeline.py updated. Grid drops to 6.5×10⁸ pixels
(73 km × 315 km bbox, matching the actual ground footprint); peak
memory 23 GB; wall time 82 s end-to-end.

Also:
- ground_to_image_batch now returns float32 (was default float64).
  Image coords are bounded by sensor dimensions (CTX ≤52k×5k,
  HiRISE ≤80k×5k); float32 represents every integer up to 16.7M
  exactly, giving sub-pixel precision <10⁻³ px anywhere in frame —
  orders of magnitude below CSM's own pointing uncertainty. Matches
  the dtype the downstream _bilinear_upsample_pair already emits,
  eliminating an implicit cast, and halves the coarse-grid
  footprint.
- Stale CoordinateMap docstring corrected (was "float64", actual
  output always was float32).
- docs/ctxpipe.qmd: new "Projection and memory notes" section
  explaining lon_0 and the float32 precision argument.
- Plans/output-tiling.md: design plan for a future tiled-output
  mode that would bound peak memory for HiRISE-scale /
  multi-strip projections where even a correctly-centered
  bounding box exceeds RAM. Benchmarks and implementation to
  follow in a dedicated session.

+++ src/isistools/csm2map/camera.py
@@ -499,8 +499,18 @@ def ground_to_image_batch(
     z = frad * sin_lat
 
     n = len(flat)
-    out_lines = np.full(n, np.nan)
-    out_samps = np.full(n, np.nan)
+    # float32 is enough for image coordinates: the line/sample values
+    # sit in 0..n_lines and 0..n_samples (CTX ≤ 52 k lines × 5 k
+    # samples, HiRISE ≤ 80 k × 5 k).  float32 represents integers up
+    # to 16_777_216 exactly; for CTX this gives >0.001-pixel precision
+    # anywhere in the frame, far below the CSM model's own pointing
+    # uncertainty (sub-pixel but many multiples of 10⁻³ px).  Using
+    # float32 also keeps the dtype consistent end-to-end: the
+    # downstream ``_bilinear_upsample_pair`` upsamples in float32
+    # already, so staying in float32 here avoids an implicit cast and
+    # halves the coarse-grid memory footprint.
+    out_lines = np.full(n, np.nan, dtype=np.float32)
+    out_samps = np.full(n, np.nan, dtype=np.float32)
 
     if workers is None:
         workers = os.cpu_count() or 1
